Fix calibrations tx on legacy comms
Note the renaming of newComms to comms. Old comms has been moved to _legacy files
