+++ reference/speeduino.ini
@@ -229,7 +229,6 @@
     pageValueWrite      = "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v"
     pageChunkWrite      = "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v", "M%2i%2o%2c%v"
     crc32CheckCommand   = "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i", "d%2i"
-    tableCrcCommand     = "k\$tsCanId%2i%2o%2c"
 
 #if mcu_stm32
     blockingFactor = 121
@@ -246,6 +245,7 @@
     restrictSquirtRelationship = false ;This requires TS 3.1 or above
 #if NEW_COMMS
     messageEnvelopeFormat = msEnvelope_1.0 ;New and testing only
+    tableCrcCommand     = "k\$tsCanId%2i%2o%2c" ;TS can only use this command in new mode
 #endif
     readSdCompressed = false
